Add movement functionality of the player
In the past, we usually used this.transform.Translate() for
change of position. However that requires us to not only handle
this aspect but also the collision aspect of things. But we already
have a rigidBody2D component attached to out object. Therefore we should
use that this component's velocity field in order to facilitate this
process for us.