Nullify parent through a public method
When a platform goes out of the camera frame, it gets destroyed.
Therefore, in order to avoid our player from getting destroyed; we
ought to have a way to nullify the parent of the player when it jumps
and the platform goes out of the frame. Otherwise there is a chance
that when the platform does get destroyed, the player is destroyed as
well; since when a parent is destroyed, all its children are destroyed
as well.