Add a public property isOnPlatform
We want to know whether or not the player has jumped on a specific
platform that we have cloned in the scene. The most straightforward
solution that I could come up with was to add a public property to
Platform class. The PlayerMovement instances could then set this
property to true or false depending on whether or not the player is
on that platform.