Add private field indexOfMarkedPlatform
We need to be able to keep track of the platform we are currently on.
In addition to that, since we are only cloning 2 platforms at a time,
we need to know the next index of clonedPlatforms[] that we need to use
to save the next clone in. Adding a field to keep track of this index
seemed to be the most sensible route to take.